Make storage access throw for meta tensors
Without throwing, we can easily segfault trying to access nullptr
storage.

To do this I made set_storage_access_should_throw public so that you
don't have to subclass TensorImpl to do it.  An alternative is
to just bite the bullet and add a MetaTensorImpl subclass.  Let
me know what is preferred.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 777a58e5650f1f8251549fdc5e4ad2744520f5ae
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53681